CI #43 polish: refund money type, lint Phase C complete, admin redirect
Post-convergence cleanup (PR #43 already fully green). Three items:

1. refund money type — `amount: float` was a false contract on the
   public refund entry points (refund_payment / refund_card_payment /
   refund_cash_payment in both payments.py and paypal_payments.py): the
   production caller (admin.refund_view) passes a quantized `Decimal`,
   and `Decimal - float` raises TypeError — the float annotation masked
   a latent crash class. Changed the 5 `amount: float` → `amount:
   Decimal` (the real monetary domain type). mypy fallout was a single
   checked test call site (test_paypal_payments.py:1424 passing int
   `10`) → `Decimal(10)`; no payment-arithmetic errors surfaced
   (existing casts hold), confirming the contract really is Decimal.
   No runtime/behaviour change — annotations only.

2. lint Phase C COMPLETE — pylint (--fail-under=9.0) and vulture are
   both green on the branch (verified on the 92ffa3a CI run), so their
   `continue-on-error: true` is removed: mypy, pylint and vulture are
   now all BLOCKING. A future type/lint/dead-code regression fails CI.

3. admin.py copy_to_event — applied the same open-redirect hardening as
   refund_view for consistency/defense-in-depth: replaced
   `HttpResponseRedirect(request.path)` with a route-reversed model
   changelist URL (`reverse("admin:<app>_<model>_changelist")`, derived
   from self.model._meta — no request-controlled data). Byte-identical
   behaviour (reverse resolves to /admin/registration/staff/, exactly
   what request.path was for this admin action). Not a CodeQL finding;
   purely consistency. No request.path redirects remain in admin.py.

Also evaluated the remaining github-advanced-security / pre-existing
CodeQL backlog (clear-text-logging, stack-trace-exposure,
url-redirection #14/#15, workflow-permissions, JS XSS): all already
remediated on this branch by the S33 work — the PR #43 head ref has
**0 open CodeQL alerts**; the `?state=open` list reflects the stale
`production` default branch and auto-closes on merge. Nothing to do.

Local-verified: mypy "Success: no issues found in 101 source files"
(blocking), ruff check + format clean, vulture clean, lint.yml valid
YAML, makemigrations "No changes detected", lint-status-writers +
lint-env-sweep clean, reverse() resolves to the prior path.

+++ registration/payments.py
@@ -488,7 +488,7 @@ def update_order_payment_data(order: Order, order_total: int, payment: Payment) 
 
 def refund_payment(
     order: Order,
-    amount: float,
+    amount: Decimal,
     reason: str | None = None,
     request: HttpRequest | None = None,
 ) -> tuple[bool, str | None]:
@@ -520,7 +520,7 @@ def refund_payment(
 
 
 def refund_cash_payment(
-    order: Order, amount: float, reason: str | None = None
+    order: Order, amount: Decimal, reason: str | None = None
 ) -> tuple[bool, None]:
     """
     Deducts the ``amount`` from the ``order``'s total and logs a `Cashdrawer`
@@ -550,7 +550,7 @@ def refund_cash_payment(
 
 def refund_card_payment(
     order: Order,
-    amount: float,
+    amount: Decimal,
     reason: str | None = None,
     request: HttpRequest | None = None,
 ) -> tuple[bool, str]:
